fix: Check for AVX instruction, and produce a warning (#20463)
When running on CPUs without an AVX instruction, the embedded MongoDB
cannot run. It fails to run with a cryptic error message, and they are
forced to reach out to support for help.

This PR adds a check, and a warning message, if trying to use embedded
MongoDB on a CPU without the AVX instruction.