fix: force auto height on no drop (#20917)
## Description
When we drag a widget within a container like widget, the container
expands instantly. However, when we don't drop the widget and bring the
mouse point outside the canvas, the auto height fails to kick in, as no
action has been performed. This results in the container like widget's
height to remain as it was. This leads to incorrect height for the
container like widget.

This PR fixes it by forcing a new computation when the container like
widget's internal canvas receives a new height.

Fixes #20907 



Media
Before


https://user-images.githubusercontent.com/103687/220900648-81cb5a0e-9aea-4d21-901e-29a8aa99c6b2.mov


After


https://user-images.githubusercontent.com/103687/220900630-725929dc-ce87-4b92-aec4-1305acd79e18.mov


## Type of change

- Bug fix (non-breaking change which fixes an issue)