chore: Check for AVX instruction earlier than the first mongod run (#21621)
We're already checking for the AVX instruction, and producing a helpful
message about it. But, we're doing that check _after_ we invoke the
`mongod` process, which fails on such processes, and the error message
is never shown.

This PR changes this, by doing the check before that `mongod` command is
run.