ci: Remove license in ci-test.yml (#21938)
The License key is unused in this repo. Setting it here is misleading,
and can cause confusions at best, or missed accidental License failures
at worst.

This PR removes the License key occurrences from `ci-test.yml`.

Note: Please review/approve only. Don't merge. This will likely cause a
conflict, so I'd like to be available when this is merged so I can
resolve the conflict. Thanks!