fix: Fix annoying issue with Entity form control in google sheets (#22103)
In google sheet query creation, when I select operation as insert many,
update one, and update many, in the entity dropdown only sheet rows
option is enabled by default, other options of spreadsheet and sheet are
disabled, Since a single option is enabled for these operations we can
default the dropdown to this option for above three operations. This PR
fixes that.