chore: Suppress sentry error capture for invalid dynamic path (#22320)
## Description
This PR suppresses sentry errors related to invalid dynamic path.

[Example](https://appsmith.sentry.io/issues/3746948945/?referrer=github_integration)

As these errors are very frequent, the quota for sentry gets filled up
very quickly. This particular error does not break any apps and was
added to sentry to gather more information as to how/when this error
occurs.

Now we have enough information to move forward in finding the solution
for this error so we are temporarily suppressing this error until we
have a fix and avoid unnecessary filling up of Sentry quota.