fix: Remove console log on MetaWidgetUpdate (#22836)
Fixes #22527 

Based on debugging the performance logs, it seems like for now that the
console log on every character type is causing a lag in the performance.

This becomes more accentuated for a list widget. Commenting out the
logging for now since the issue is marked critical

In the meanwhile, @SatishGandham will help to triage the further cause
of this issue.