fix: Return default page id instead of id in restore snapshot API (#23096)
## Description
The restore snapshot API returns the Application object that was
restored. In current implementation, it returns pages with `_id` as the
`id`. This creates problem when the Application is Git connected because
no page will be found with the _id and branch name. This PR fixes this
problem by replacing the `id` with the `defaultPageId`.

#### PR fixes following issue(s)
Fixes #22956